ARM NEON support plus fix a number of warnings.

+++ OgreMain/src/Compositor/OgreCompositorNodeDef.cpp
@@ -61,7 +61,7 @@ namespace Ogre
 
 		if( textureSource != TEXTURE_GLOBAL )
 		{
-			mOutChannelMapping[outChannel] = encodeTexSource( index, textureSource );
+			mOutChannelMapping[outChannel] = encodeTexSource( (uint32)index, textureSource );
 		}
 		else
 		{
